add new logo and reporting page header
