+++ app/src/main/shared/components/PageWithHeader/Page.tsx
@@ -1,21 +1,18 @@
 import * as React from "react";
-import { PageHeader } from "./PageHeader";
-import { PageArticle } from "./PageArticle";
-import { PageParts } from "./PageWithHeader";
+import {PageArticle} from "./PageArticle";
+import {PageParts} from "./PageWithHeader";
 
 interface PageProps {
     page: PageParts;
 }
 
-export class Page extends React.Component<PageProps, undefined> {
+export class PageNoHeader extends React.Component<PageProps, undefined> {
 
-    render() :JSX.Element {
+    render(): JSX.Element {
         const page = this.props.page;
-        return <div>
-            <PageHeader siteTitle={page.siteTitle()} />
-            <PageArticle title={page.title()} hideTitle={page.hideTitle()}>
-                { this.props.children }
-            </PageArticle>
-         </div>;
+
+        return <PageArticle title={page.title()} hideTitle={page.hideTitle()}>
+            {this.props.children}
+        </PageArticle>
     }
 }
